+++ resnet/src/functional/add.cu
@@ -1,11 +1,11 @@
 /** @file add.cu
 */
 
-#include "functional/macros.h"
+#include "functional/macros.hpp"
 #include "functional/add.hpp"
-#include "common.h"
+#include "common.hpp"
 
-__global__ static void cuda_add_(float *Result, const float *adder, int length) {
+__global__ static void cuda_add_(float *Result, const float *adder, size_t length) {
   CUDA_KERNEL_LOOP(i, length) {
     if (i < length) {
       Result[i] += adder[i];
@@ -13,7 +13,7 @@ __global__ static void cuda_add_(float *Result, const float *adder, int length) 
   }
 }
 
-__global__ static void cuda_add(float *Result, const float *adder_a, const float *adder_b, int length) {
+__global__ static void cuda_add(float *Result, const float *adder_a, const float *adder_b, size_t length) {
   CUDA_KERNEL_LOOP(i, length) {
     if (i < length) {
       Result[i] = adder_a[i] + adder_b[i];
@@ -21,7 +21,7 @@ __global__ static void cuda_add(float *Result, const float *adder_a, const float
   }
 }
 
-__global__ static void cuda_relu_(float *Result, int length) {
+__global__ static void cuda_relu_(float *Result, size_t length) {
   CUDA_KERNEL_LOOP(i, length) {
     if (i < length) {
       Result[i] = Result[i] > 0.0f ? Result[i] : 0.0f;
@@ -29,7 +29,7 @@ __global__ static void cuda_relu_(float *Result, int length) {
   }
 }
 
-__global__ static void cuda_add_relu_(float *Result, const float *adder, int length) {
+__global__ static void cuda_add_relu_(float *Result, const float *adder, size_t length) {
   CUDA_KERNEL_LOOP(i, length) {
     if (i < length) {
       Result[i] += adder[i];
@@ -38,16 +38,18 @@ __global__ static void cuda_add_relu_(float *Result, const float *adder, int len
   }
 }
 
-
-void add_(float *Result, const float *adder, int length, Impl::DeviceType device_type, cudaStream_t stream) {
+void add_(float *Result, const float *adder, size_t length, Impl::DeviceType device_type, cudaStream_t stream) {
   switch (device_type) {
   case Impl::DeviceType::CPU: {
-    for (int i = 0; i < length; ++i) {
+    for (size_t i = 0; i < length; ++i) {
       Result[i] += adder[i];
     }
   }
     break;
-  case Impl::DeviceType::CUDA:cuda_add_<<<KERNEL_LOOP_BLOCKS(length), KERNEL_LOOP_THREADS, 0, stream>>>(Result, adder, length);
+  case Impl::DeviceType::CUDA:
+    cuda_add_<<<KERNEL_LOOP_BLOCKS(length), KERNEL_LOOP_THREADS, 0, stream>>>(Result,
+                                                                              adder,
+                                                                              length);
     break;
   }
 }
@@ -69,10 +71,10 @@ void add(float *Result, const float *adder_a, const float *adder_b, int length, 
   }
 }
 
-void relu_(float *Result, int length, Impl::DeviceType device_type) {
+void relu_(float *Result, size_t length, Impl::DeviceType device_type) {
   switch (device_type) {
   case Impl::DeviceType::CPU: {
-    for (int i = 0; i < length; ++i) {
+    for (size_t i = 0; i < length; ++i) {
       Result[i] = Result[i] > 0.0f ? Result[i] : 0.0f;
     }
   }
@@ -82,16 +84,19 @@ void relu_(float *Result, int length, Impl::DeviceType device_type) {
   }
 }
 
-void add_relu_(float *Result, const float *adder, int length, Impl::DeviceType device_type) {
+void add_relu_(float *Result, const float *adder, size_t length, Impl::DeviceType device_type) {
   switch (device_type) {
   case Impl::DeviceType::CPU: {
-    for (int i = 0; i < length; ++i) {
+    for (size_t i = 0; i < length; ++i) {
       Result[i] += adder[i];
       Result[i] = Result[i] > 0.0f ? Result[i] : 0.0f;
     }
   }
     break;
-  case Impl::DeviceType::CUDA:cuda_add_relu_<<<KERNEL_LOOP_BLOCKS(length), KERNEL_LOOP_THREADS>>>(Result, adder, length);
+  case Impl::DeviceType::CUDA:
+    cuda_add_relu_<<<KERNEL_LOOP_BLOCKS(length), KERNEL_LOOP_THREADS>>>(Result,
+                                                                        adder,
+                                                                        length);
     break;
   }
 }
